FIX: change the sut_version field to not_analyzed
By default ElasticSearch uses analyzed string fields, however this means that the string is
broken down in smaller pieces which may or may not match a search for the exact value (as
explained in the DSL query documentation). This can cause, for example, that a search for
6.3.0-AB1 to not match. This patch changes the mapping in order to set the sut_field as
not_analyzed to prevent this.